# server/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beskar API",
    "version": "1.0.0",
    "description": "Authentication and implemented whiteboarding v2 APIs.\n\n**Try it:** [Sign in](/auth/login?returnTo=%2Fapi%2Fdocs), then use the browser-session option, or select BearerToken and paste a valid Zitadel JWT (without the Bearer prefix). The server currently verifies bearer JWTs using its OIDC ID-token verifier; opaque access tokens are not supported. Calls run against this origin with your permissions and can create real data.\n\nReplace example space/parent IDs with your own. Generate an Idempotency-Key UUID for each new board or checkpoint batch and keep it with the same payload for retries. Login, registration and logout require browser navigation. Cookie sessions normally require HTTPS.",
    "license": {
      "name": "Apache 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This Beskar instance"
    }
  ],
...
  "security": [
    {
      "BrowserSession": []
    },
    {
      "BearerToken": []
    }
  ],
  "paths": {
...
    "/api/v2/editor/space/{spaceId}/whiteboard/{pageId}/publish": {
      "post": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "publishWhiteboardV2",
        "summary": "Publish a whiteboard at a durable sequence",
        "description": "Requires edit permission. Reconstructs Yjs state through sequence and atomically creates its snapshot, version, and current publication pointer. Retries return the original version without moving the pointer. Keys are scoped to board and actor. Archived spaces reject new publications. Replay is bounded to 32 MiB and 10,000 updates. Requires whiteboard_publication.xml and the pinned Node/Yjs runtime. Title edits use optional checkpoint title metadata; the requested sequence determines the published title. Deploy whiteboard_titles.xml. Asset integration is deferred; asset records are preserved without validation or retention. Requires a client-generated PNG in preview.contentType/data. PNG is validated and stored atomically with publication; retries must keep the same sequence and image. 6 MiB JSON body limit. Deploy whiteboard_previews.xml. The server does not verify visual correspondence to the Yjs snapshot.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "example": 42,
            "description": "Existing v2 whiteboard page ID in the supplied space."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "description": "Exactly one nonzero UUID per batch. Keep the same key and decoded bytes until acknowledged. Scoped by board and authenticated actor. Different content with the same scoped key returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishWhiteboardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Immutable published manifest.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardPublishedResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST or INVALID_PREVIEW",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "403": {
            "description": "WHITEBOARD_FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "404": {
            "description": "WHITEBOARD_NOT_FOUND or WHITEBOARD_NOT_PUBLISHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "500": {
            "description": "WHITEBOARD_PUBLICATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "409": {
            "description": "IDEMPOTENCY_KEY_REUSED, SPACE_ARCHIVED, PUBLISH_SEQUENCE_UNAVAILABLE, PUBLISH_STATE_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "413": {
            "description": "REQUEST_TOO_LARGE or PUBLISH_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "503": {
            "description": "WHITEBOARD_PUBLISH_BUSY; Retry-After: 1",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/editor/space/{spaceId}/whiteboard/{pageId}/published": {
      "get": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "getPublishedWhiteboardV2",
        "summary": "Get the current published whiteboard",
        "description": "Requires view permission. Returns an immutable snapshot download URL. Archived spaces remain readable. No replay lease. preview=true returns only the stored PNG, with current view permission; preview=false returns the manifest. Returns WHITEBOARD_NOT_PUBLISHED before first publication.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "example": 42,
            "description": "Existing v2 whiteboard page ID in the supplied space."
          },
          {
            "name": "preview",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "true returns raw PNG only; false or omitted returns the manifest."
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable published manifest.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardPublishedResponse"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "403": {
            "description": "WHITEBOARD_FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "404": {
            "description": "WHITEBOARD_NOT_FOUND or WHITEBOARD_NOT_PUBLISHED or WHITEBOARD_PREVIEW_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "500": {
            "description": "WHITEBOARD_PUBLICATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "headCurrentPublishedWhiteboardV2",
        "summary": "Inspect the current published representation",
        "description": "Requires view permission. Returns an immutable snapshot download URL. Archived spaces remain readable. No replay lease. preview=true returns only the stored PNG, with current view permission; preview=false returns the manifest. Returns WHITEBOARD_NOT_PUBLISHED before first publication.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "example": 42,
            "description": "Existing v2 whiteboard page ID in the supplied space."
          },
          {
            "name": "preview",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "true returns raw PNG only; false or omitted returns the manifest."
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable published manifest.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardPublishedResponse"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "403": {
            "description": "WHITEBOARD_FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "404": {
            "description": "WHITEBOARD_NOT_FOUND or WHITEBOARD_NOT_PUBLISHED or WHITEBOARD_PREVIEW_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "500": {
            "description": "WHITEBOARD_PUBLICATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/editor/space/{spaceId}/whiteboard/{pageId}/published/{versionId}/content": {
      "get": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "downloadPublishedWhiteboardV2",
        "summary": "Download an immutable published snapshot",
        "description": "Requires current view permission and matching board/space/version. Historical published versions remain accessible. No query parameters. Supports one byte range, ETag and If-Range; HEAD ignores Range. Cache-Control: no-store.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "example": 42,
            "description": "Existing v2 whiteboard page ID in the supplied space."
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Yjs update-v1 full snapshot.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "Requested byte range."
          },
          "304": {
            "description": "ETag matched."
          },
          "416": {
            "description": "Invalid or multiple ranges."
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "403": {
            "description": "WHITEBOARD_FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "404": {
            "description": "WHITEBOARD_NOT_FOUND or WHITEBOARD_NOT_PUBLISHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "500": {
            "description": "WHITEBOARD_PUBLICATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "headPublishedWhiteboardV2",
        "summary": "Inspect an immutable published snapshot",
        "description": "Requires current view permission and matching board/space/version. Historical published versions remain accessible. No query parameters. Supports one byte range, ETag and If-Range; HEAD ignores Range. Cache-Control: no-store.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "example": 42,
            "description": "Existing v2 whiteboard page ID in the supplied space."
          },
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Yjs update-v1 full snapshot.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "Requested byte range."
          },
          "304": {
            "description": "ETag matched."
          },
          "416": {
            "description": "Invalid or multiple ranges."
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "403": {
            "description": "WHITEBOARD_FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "404": {
            "description": "WHITEBOARD_NOT_FOUND or WHITEBOARD_NOT_PUBLISHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          },
          "500": {
            "description": "WHITEBOARD_PUBLICATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Valid Zitadel JWT accepted by the server OIDC verifier. Paste only the token."
      },
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zitadel.session",
        "description": "Sign in using the link above. The browser sends the HttpOnly session cookie automatically; do not paste a cookie value."
      }
    },
    "schemas": {
...
      "WhiteboardError": {
        "type": "object",
        "required": [
          "status",
          "error"
        ],
        "properties": {
          "status": {
            "const": "FAILED"
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "WhiteboardPublishedManifest": {
        "type": "object",
        "required": [
          "pageId",
          "spaceId",
          "versionId",
          "versionNumber",
          "publishedBy",
          "publishedAt",
          "snapshot",
          "preview"
        ],
        "properties": {
          "pageId": {
            "type": "integer",
            "format": "int64"
          },
          "spaceId": {
            "type": "string",
            "format": "uuid"
          },
          "versionId": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "type": "string",
            "pattern": "^[1-9][0-9]*$"
          },
          "publishedBy": {
            "type": "string",
            "format": "uuid"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "snapshot": {
            "$ref": "#/components/schemas/WhiteboardDraftSnapshot"
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WhiteboardPreviewMetadata"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "WhiteboardPublishedResponse": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/WhiteboardPublishedManifest"
          }
        }
      },
      "PublishWhiteboardRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sequence",
          "preview"
        ],
        "properties": {
          "sequence": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$",
            "description": "Requested durable boundary; at most 9223372036854775807.",
            "example": "103"
          },
          "preview": {
            "$ref": "#/components/schemas/WhiteboardPublishPreview"
          }
        }
      },
      "WhiteboardPublishPreview": {
        "type": "object",
        "required": [
          "contentType",
          "data"
        ],
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/png"
            ]
          },
          "data": {
            "type": "string",
            "contentEncoding": "base64",
            "description": "PNG bytes without data URL prefix; max 4 MiB, 4096 pixels per dimension, 8 million pixels total."
          }
        }
      },
      "WhiteboardPreviewMetadata": {
        "type": "object",
        "required": [
          "url",
          "contentType",
          "byteLength",
          "digest",
          "width",
          "height"
        ],
        "properties": {
          "url": {
            "type": "string"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/png"
            ]
          },
          "byteLength": {
            "type": "integer"
          },
          "digest": {
            "type": "string"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}